+++ data/korean.txt
