+++ setup.py
@@ -1,0 +1,6 @@
+from distutils.core import setup
+from setuptools import find_packages
+
+setup(name='TajaBot',
+      version='0.1',
+      packages=find_packages(include=['taja']))
